requirements.txt
selenium
beautifulsoup4
bs4
webdriver_manager
requests
pandas
fastapi
uvicorn
python-dotenv
langdetect
nltk
google-api-python-client